Add ability to configure an additional https port
This is needed because we have 2 different certificates we need to be
able to use, and we have client that needs to use a specific one of them
that cannot be referenced by SNI after Java 17 upgrade because of
https://bugs.openjdk.org/browse/JDK-8261969.

TODO: remove custom version changes, finish updating the TestHttpServerConfig tests, write
tests.

This commit is based on release 0.216.

Will need this as a hotfix on 0.216 and a new commit/release off of master
that we can use in the jetty upgrade pr and after. There are slight
changes needed between the two due to the java upgrade.